Make Fibonacci benchmarks behave a little better
I'm not sure anyone really wants to try to guess how relevant
a benchmark is based on fib(b.N), so this adjusts the benchmark to
simply call Fibonacci functions repeatedly with a default N=20. N can
be adjusted by setting the env var BENCH_LUA_FIB_N to an integer >= 0.